Remove bug with hint constantly rendering and move DeviceLocation interface to appropriate folder
